Fix test module import, extract gate types, refresh length budget
- AgentNotificationGateTests imported only `cmux`; the app-target types
  live in the `cmux_DEV` test-host module, so the bundle failed to compile
  and every unit-test shard failed. Import both, matching every other test.
- Move AgentNotifyCategory / AgentTurnCompleteMode / AgentNotificationMeta /
  AgentNotificationGate out of the 14k-line TerminalController.swift into
  Sources/AgentNotificationGate.swift (pure, better isolated, shrinks the
  file's growth). Wired into the app target.
- Refresh .github/swift-file-length-budget.tsv for the three files that grew
  (CLI/cmux.swift, TerminalController.swift, AppSection.swift).
